move callbackSerializer to grpcsync package
